+++ src/parser.c
@@ -3,7 +3,7 @@
 // TODO provare ad aggiungere 1 alla lunghezza delle stringe generate dalle funzioni
 
 // Takes as input the parameter to search, the file in which the parameter is to be searched and the variable in which to save its value
-int parse(char* path, char* key, unsigned* dest) {
+int parse(char* path, char* key, unsigned long* dest) {
   int tempErr;     // Variable used for temporary error codes
   int source;     // Declaration of File Descriptor
   if((source=open(path, RPERM, NULL))==ERR) {     // Assigning config file to the FD source + elementary error handling
